% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackBill.R
\name{blackBilldeleteServer}
\alias{blackBilldeleteServer}
\title{处理逻辑}
\usage{
blackBilldeleteServer(input, output, session, erp_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{erp_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
blackBilldeleteServer()
}
